Move check config loading to own package
